f64ification

+++ src/point/nalgebra.rs
@@ -3,31 +3,31 @@
 use super::Point;
 
 impl Point for nalgebra::Vector2<f32> {
-	fn x(&self) -> f32 {
-		self.x
+	fn x(&self) -> f64 {
+		self.x as f64
 	}
 
-	fn y(&self) -> f32 {
-		self.y
+	fn y(&self) -> f64 {
+		self.y as f64
 	}
 
-	fn edit(&mut self, x: f32, y: f32) {
-		self.x = x;
-		self.y = y;
+	fn edit(&mut self, x: f64, y: f64) {
+		self.x = x as f64;
+		self.y = y as f64;
 	}
 }
 
 impl Point for nalgebra::Vector2<f64> {
-	fn x(&self) -> f32 {
-		self.x as f32
+	fn x(&self) -> f64 {
+		self.x
 	}
 
-	fn y(&self) -> f32 {
-		self.y as f32
+	fn y(&self) -> f64 {
+		self.y
 	}
 
-	fn edit(&mut self, x: f32, y: f32) {
-		self.x = x as f64;
-		self.y = y as f64;
+	fn edit(&mut self, x: f64, y: f64) {
+		self.x = x;
+		self.y = y;
 	}
 }
